refactor(inference): page index is an uint8_t

+++ apps/inference/models/statistic/one_mean_statistic.cpp
@@ -3,7 +3,7 @@
 namespace Inference {
 
 void OneMeanStatistic::syncParametersWithStore(const Statistic* stat) {
-  if (!hasSeries(static_cast<int>(m_activePageIndex))) {
+  if (!hasSeries(m_activePageIndex)) {
     return;
   }
   m_params[OneMean::ParamsOrder::x] = mean(series());
